Move SleepSchedule from workers to main bot instance and some whitespace cleanup
