+++ pokemongo_bot/cell_workers/__init__.py
@@ -17,7 +17,6 @@ from follow_path import FollowPath
 from follow_spiral import FollowSpiral
 from collect_level_up_reward import CollectLevelUpReward
 from follow_cluster import FollowCluster
-from sleep_schedule import SleepSchedule
 from update_live_stats import UpdateLiveStats
 from update_live_inventory import UpdateLiveInventory
 from catch_pokemon import CatchPokemon
